Added control for valid target

+++ src/Idephix/Extension/Project/Project.php
@@ -13,6 +13,9 @@ use Idephix\Extension\IdephixAwareInterface;
  */
 class Project implements IdephixAwareInterface
 {
+    /**
+     * @var \Idephix\IdephixInterface
+     */
     private $idx;
 
     public function setIdephix(IdephixInterface $idx)
@@ -27,15 +30,20 @@ class Project implements IdephixAwareInterface
         }
 
         $target = $this->idx->getCurrentTarget();
-        $user = $target->get('ssh_params.user');
-        $host = $this->idx->getCurrentTargetHost();
 
-        if (file_exists($exclude)) {
-          $extraOpts .= ' --exclude-from='.$exclude;
+        if( $target !== null ) {
+            $user = $target->get('ssh_params.user');
+            $host = $this->idx->getCurrentTargetHost();
+
+            if (file_exists($exclude)) {
+                $extraOpts .= ' --exclude-from='.$exclude;
+            }
+
+            $cmd = "rsync -rlDcz --force --delete --progress $extraOpts -e 'ssh' $localDir $user@$host:$remoteDir";
+
+            return $this->idx->local($cmd);
+        } else {
+            $this->idx->output->writeln("<error>Target not provided. Please provide a valid target.</error>");
         }
-
-        $cmd = "rsync -rlDcz --force --delete --progress $extraOpts -e 'ssh' $localDir $user@$host:$remoteDir";
-
-        return $this->idx->local($cmd);
     }
 }
